+++ CHANGELOG.md
@@ -5,13 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased](https://github.com/cockroachdb/cockroach-operator/compare/v2.6.0...master)
+# [Unreleased](https://github.com/cockroachdb/cockroach-operator/compare/v2.7.0...master)
+
+# [v2.7.0](https://github.com/cockroachdb/cockroach-operator/compare/v2.6.0...v2.7.0)
 
 ## Fixed
 
 * Grant operator deletecollection permissions to fix fullcluster restart flow
 * Grant operator list and update permissions on pvcs to fix pvc resize flow
 * Bump TerminationGracePeriodSeconds from 1m to 5m
+* Prefer user added --join flags over default when explicitly passed
 
 # [v2.6.0](https://github.com/cockroachdb/cockroach-operator/compare/v2.5.3...v2.6.0)